requirements.txt
anthropic>=0.28.0
gspread>=6.0.0
google-auth>=2.28.0
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
